refact(new-price-calculator): promote purchase summary to a separate component instead of a part of the purchase form
