+++ apps/store/src/features/priceCalculator/PurchaseFormV2/PurchaseFormV2.css.ts
@@ -26,18 +26,3 @@ export const priceLoaderWrapper = style([
 ])
 
 export const viewOffersWrapper = style([yStack(), centered, { gap: '2.75rem' }])
-
-export const purchaseSummaryWrapper = style([
-  yStack({ alignItems: 'center', justifyContent: 'center' }),
-  { height: '100%' },
-])
-
-export const purchaseSummary = style({
-  width: `min(100%, ${CONTENT_MAX_WIDTH})`,
-  marginInline: 'auto',
-  ...responsiveStyles({
-    lg: {
-      width: `max(100%, ${CONTENT_MAX_WIDTH})`,
-    },
-  }),
-})
